# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Noodle Engine API",
    "version": "1.0.0a10",
    "description": "Drive the Noodle test engine from another system: every MCP tool as a plain JSON endpoint (NOOD_0193), for callers that can't speak MCP.\n\nThis is the ENGINE API \u2014 the caller side. It is not the 'api wok', which is Noodle testing someone else's REST service with @api feature files; nothing here is a test target.\n\nCalls are SYNCHRONOUS: a run returns when it finishes, so set client read timeouts in minutes, not seconds. One call runs at a time per server. Green means `failed == 0` AND `verified == true`."
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "noodle-mcp --transport streamable-http"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    },
    {
      "apiKeyAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      },
      "apiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key"
      }
    }
  },
...
  "paths": {
...
    "/api/tools/generate_test": {
      "post": {
        "summary": "Create a Gherkin .feature file + page-object skeleton from a plain-English",
        "description": "Create a Gherkin .feature file + page-object skeleton from a plain-English\ndescription of a web test. Rule-based templates by default (login / search /\ncheckbox / dropdown / generic, with quoted values slot-filled); use_llm=True\nroutes generation through the workspace's configured NOODLE_MODEL instead.\nA bare @tag token in `description` (e.g. \"tag this @smoke\") is added to\nthe generated scenario(s). append_to (an existing feature's file stem,\ne.g. \"search_suggestion\") adds this test case's scenario(s) to that file\ninstead of writing a separate one \u2014 same app/topic, one more scenario;\nomit it to always get a new .feature file, e.g. for a different suite or\ntopic (rule-based generation only; ignored with use_llm). Returns the\nwritten file paths, whether the test is runnable as-is (runnable=False\nmeans <placeholders> still need real values), and the generator's log\noutput. workspace overrides the server's default workspace for this call.",
        "operationId": "generate_test",
        "tags": [
          "tools"
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "url": {
                    "title": "Url",
                    "type": "string"
                  },
                  "description": {
                    "title": "Description",
                    "type": "string"
                  },
                  "use_llm": {
                    "default": false,
                    "title": "Use Llm",
                    "type": "boolean"
                  },
                  "overwrite": {
                    "default": false,
                    "title": "Overwrite",
                    "type": "boolean"
                  },
                  "append_to": {
                    "anyOf": [
                      {
                        "type": "string"
                      },
                      {
                        "type": "null"
                      }
                    ],
                    "default": null,
                    "title": "Append To"
                  },
                  "workspace": {
                    "anyOf": [
                      {
                        "type": "string"
                      },
                      {
                        "type": "null"
                      }
                    ],
                    "default": null,
                    "title": "Workspace"
                  }
                },
                "required": [
                  "url",
                  "description"
                ],
                "title": "generate_testArguments",
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "the tool's payload",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "malformed JSON body"
          },
          "401": {
            "description": "missing or wrong API key"
          },
          "404": {
            "description": "no such tool"
          },
          "500": {
            "description": "the tool itself failed; `error` names the cause"
          }
        }
      }
    },
...
    "/api/tools/run_and_report": {
      "post": {
        "summary": "THE execution operation: preflight \u2192 run \u2192 report \u2192 serve, one bounded",
        "description": "THE execution operation: preflight \u2192 run \u2192 report \u2192 serve, one bounded\npayload. A missing/placeholder {env:KEY} returns missing_secret_keys with\nZERO browser launches. The run hook's reports are freshness-checked\nonce (rebuilt only if stale); serve_reports defaults True \u2014 the workflow\nends on served Allure+RCA URLs, never file paths. On red the\npayload carries rca_compact (verdict + failing step + fix): read that, do\nnot call get_rca separately; on a locator/state failure reproduce that\nexact state ONCE (probe_page do=[the flow]) and re-author from evidence\n\u2014 never one guessed fix per red run. Same target and\nheadless/browser/retries/parallel flags as run_test (pass headless=True,\nretries=0 while iterating). workspace overrides the server's default.",
        "operationId": "run_and_report",
        "tags": [
          "tools"
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "target": {
                    "anyOf": [
                      {
                        "type": "string"
                      },
                      {
                        "type": "null"
                      }
                    ],
                    "default": null,
                    "title": "Target"
                  },
                  "tag": {
                    "anyOf": [
                      {
                        "type": "string"
                      },
                      {
                        "type": "null"
                      }
                    ],
                    "default": null,
                    "title": "Tag"
                  },
                  "workspace": {
                    "anyOf": [
                      {
                        "type": "string"
                      },
                      {
                        "type": "null"
                      }
                    ],
                    "default": null,
                    "title": "Workspace"
                  },
                  "headless": {
                    "anyOf": [
                      {
                        "type": "boolean"
                      },
                      {
                        "type": "null"
                      }
                    ],
                    "default": null,
                    "title": "Headless"
                  },
                  "browser": {
                    "anyOf": [
                      {
                        "type": "string"
                      },
                      {
                        "type": "null"
                      }
                    ],
                    "default": null,
                    "title": "Browser"
                  },
                  "retries": {
                    "anyOf": [
                      {
                        "type": "integer"
                      },
                      {
                        "type": "null"
                      }
                    ],
                    "default": null,
                    "title": "Retries"
                  },
                  "parallel": {
                    "anyOf": [
                      {
                        "type": "integer"
                      },
                      {
                        "type": "null"
                      }
                    ],
                    "default": null,
                    "title": "Parallel"
                  },
                  "parallel_scheme": {
                    "default": "feature",
                    "title": "Parallel Scheme",
                    "type": "string"
                  },
                  "serve_reports": {
                    "default": true,
                    "title": "Serve Reports",
                    "type": "boolean"
                  }
                },
                "title": "run_and_reportArguments",
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "the tool's payload",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "malformed JSON body"
          },
          "401": {
            "description": "missing or wrong API key"
          },
          "404": {
            "description": "no such tool"
          },
          "500": {
            "description": "the tool itself failed; `error` names the cause"
          }
        }
      }
    },
...
    "/api/tools/author_test": {
      "post": {
        "summary": "Write a whole test package in ONE transaction (web/<app_name> or",
        "description": "Write a whole test package in ONE transaction (web/<app_name> or\nthe existing package mapped to base_url \u2014 environments.yaml, POM,\nfeature): validated first; any failure rolls every byte back.\nfeature_content / pom_content are each ONE string.\nsecret_values ({KEY: value}) is WRITE-ONLY (gitignored\n<app>_secrets.env, never echoed); missing required_secret_keys become\nplaceholders. `prompt`: numbered plain-English steps, expanded\nDETERMINISTICALLY to a goal (underspecified steps borrow from\nneighbours, inferences echoed as prompt_expansion.assumptions; a\nURL derives the three path params).\n\nELSE prefer `goal` (mutually exclusive w/ feature_content): {scenario,\nnavigation?: [ordered URLs, an {env:} Given each],\nactions: [{do: search|suggest|pick|add_to|click|enter|select|check|\nuncheck|hover|upload|press_key|pick_date|go_back, id?, term/target/\nvalue/option/file/key/date; pick binds 'any matching result' to ONE\nprobed result item; add_to = {item_from: pick-id, destination},\nengine-lowered to observed controls; never invent surface steps}],\nchecks: [{see|not_see|count|any_of|field|item_in_destination|\nurl_contains, min?, value?,\nexpected_from?: pick-id, evidence?: screenshot, after?: action-id}],\ndismissals?, probe?: {discover}}. item_in_destination re-asserts the\nbound caption (identity, not count). ONE goal-scoped probe,\nengine-compiled feature + POM; unproven requests block, never drop.\nPayloads carry goal_verified, intent_verified + intent_trace; a block\nnames ONE typed next_action \u2014 repair that gap only. Manual\nfeature_content is never intent_verified; allow_unverified_intent is\na human-only override.\nrun_after_author=true also runs once (headless, retries=0), serves\nboth reports, returns {author, run}; 0 passed = failure.\n\nReturns paths, warnings, missing_secret_keys, base_url_key (use that\nexact {env:} key), and ready+blocking. ready=true = static authoring\nchecks \u2014 never validate/preflight separately; run next. ready=false:\nfix `blocking`, re-author with overwrite=true. workspace overrides\nthe server default.",
        "operationId": "author_test",
        "tags": [
          "tools"
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "app_name": {
                    "anyOf": [
                      {
                        "type": "string"
                      },
                      {
                        "type": "null"
                      }
                    ],
                    "default": null,
                    "title": "App Name"
                  },
                  "base_url": {
                    "anyOf": [
                      {
                        "type": "string"
                      },
                      {
                        "type": "null"
                      }
                    ],
                    "default": null,
                    "title": "Base Url"
                  },
                  "feature_path": {
                    "anyOf": [
                      {
                        "type": "string"
                      },
                      {
                        "type": "null"
                      }
                    ],
                    "default": null,
                    "title": "Feature Path"
                  },
                  "feature_content": {
                    "anyOf": [
                      {
                        "type": "string"
                      },
                      {
                        "type": "null"
                      }
                    ],
                    "default": null,
                    "title": "Feature Content"
                  },
                  "pom_content": {
                    "anyOf": [
                      {
                        "type": "string"
                      },
                      {
                        "type": "null"
                      }
                    ],
                    "default": null,
                    "title": "Pom Content"
                  },
                  "environment_values": {
                    "anyOf": [
                      {
                        "additionalProperties": true,
                        "type": "object"
                      },
                      {
                        "type": "null"
                      }
                    ],
                    "default": null,
                    "title": "Environment Values"
                  },
                  "required_secret_keys": {
                    "anyOf": [
                      {
                        "items": {
                          "type": "string"
                        },
                        "type": "array"
                      },
                      {
                        "type": "null"
                      }
                    ],
                    "default": null,
                    "title": "Required Secret Keys"
                  },
                  "secret_values": {
                    "anyOf": [
                      {
                        "additionalProperties": true,
                        "type": "object"
                      },
                      {
                        "type": "null"
                      }
                    ],
                    "default": null,
                    "title": "Secret Values"
                  },
                  "goal": {
                    "anyOf": [
                      {
                        "additionalProperties": true,
                        "type": "object"
                      },
                      {
                        "type": "null"
                      }
                    ],
                    "default": null,
                    "title": "Goal"
                  },
                  "run_after_author": {
                    "default": false,
                    "title": "Run After Author",
                    "type": "boolean"
                  },
                  "overwrite": {
                    "default": false,
                    "title": "Overwrite",
                    "type": "boolean"
                  },
                  "allow_unverified_intent": {
                    "default": false,
                    "title": "Allow Unverified Intent",
                    "type": "boolean"
                  },
                  "prompt": {
                    "anyOf": [
                      {
                        "type": "string"
                      },
                      {
                        "type": "null"
                      }
                    ],
                    "default": null,
                    "title": "Prompt"
                  },
                  "workspace": {
                    "anyOf": [
                      {
                        "type": "string"
                      },
                      {
                        "type": "null"
                      }
                    ],
                    "default": null,
                    "title": "Workspace"
                  }
                },
                "title": "author_testArguments",
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "the tool's payload",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "malformed JSON body"
          },
          "401": {
            "description": "missing or wrong API key"
          },
          "404": {
            "description": "no such tool"
          },
          "500": {
            "description": "the tool itself failed; `error` names the cause"
          }
        }
      }
    },
...
    "/api/tools/write_feature": {
      "post": {
        "summary": "Write caller-authored Gherkin into the workspace (path must be under",
        "description": "Write caller-authored Gherkin into the workspace (path must be under\nthe tests dir and end in .feature). Content is validated as Gherkin first;\nsteps that would need an LLM fallback at runtime are reported back.\nworkspace overrides the server's default workspace for this call.",
        "operationId": "write_feature",
        "tags": [
          "tools"
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "path": {
                    "title": "Path",
                    "type": "string"
                  },
                  "content": {
                    "title": "Content",
                    "type": "string"
                  },
                  "overwrite": {
                    "default": false,
                    "title": "Overwrite",
                    "type": "boolean"
                  },
                  "workspace": {
                    "anyOf": [
                      {
                        "type": "string"
                      },
                      {
                        "type": "null"
                      }
                    ],
                    "default": null,
                    "title": "Workspace"
                  }
                },
                "required": [
                  "path",
                  "content"
                ],
                "title": "write_featureArguments",
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "the tool's payload",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "malformed JSON body"
          },
          "401": {
            "description": "missing or wrong API key"
          },
          "404": {
            "description": "no such tool"
          },
          "500": {
            "description": "the tool itself failed; `error` names the cause"
          }
        }
      }
    },
...
    "/api/tools/get_rca": {
      "post": {
        "summary": "Root-cause analysis of every failed/errored scenario from the last",
        "description": "Root-cause analysis of every failed/errored scenario from the last\nrun. compact=True (default, NOOD_0117) is the cheap-evidence-first read:\nverdict + failing step + suggested fix per failure, a few lines total \u2014\nusually all a fix needs; read it BEFORE any screenshot (vision tokens)\nor network capture. compact=False returns the full markdown report.\nworkspace overrides the server's default workspace for this call.",
        "operationId": "get_rca",
        "tags": [
          "tools"
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "workspace": {
                    "anyOf": [
                      {
                        "type": "string"
                      },
                      {
                        "type": "null"
                      }
                    ],
                    "default": null,
                    "title": "Workspace"
                  },
                  "compact": {
                    "default": true,
                    "title": "Compact",
                    "type": "boolean"
                  }
                },
                "title": "get_rcaArguments",
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "the tool's payload",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "malformed JSON body"
          },
          "401": {
            "description": "missing or wrong API key"
          },
          "404": {
            "description": "no such tool"
          },
          "500": {
            "description": "the tool itself failed; `error` names the cause"
          }
        }
      }
    },
...
  }
}